backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1
